requirements.txt
alembic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
kafka-python
psycopg2